Was broken on radio elements too, not just select's on their data tab. Looks like it's broken for all components that use datagrid to list potential options in the editor popup. So changed the code to check for builder and the problem, it otherwise operates as designed by lib author. The radio component apparently shoves non-objects into the array as well.
I am left only to presume it works as intended for him in some way at some point in time in some location of the program, and that he just isn't debugging or checking his work on the builder implementation at this time.
Not trying to be rude, just a clear observation. The builder needs some seroius attention
#553