forcing the component changed check for advanced logic impl to utilize the reducer's resultant value, instead of just a lodash equality comparer on the objects. Since it's always different (original comp clone (aka new comp) cannot ever be same as current in-memory comp), resulting in an infinite loop.
Additionally switched the call to use isMatch since the deep clone is based upon the ORIGINAL component, not on the ACTUAL component in memory, and we need a loose comparison against defined props in the new comp.
Additionally, the action reducer wasn't responding properly with a [true] value when the logic action type is a "property".

Sadly the formio system doesn't currently produce anti-logic resultants. So if a previous logic triggered an action, the reversal of the change is not executed.
I might revisit this to avoid having to recreate duplicate logic to reverse the change made.
The original code would overwrite the current component entirely, which would've applied the reversal. But it would be an infinite loop without the isMatch call, which cannot detect that change reliably without a customizer function to handle the depth scan manually.